Update README.md
Using the dataset, we constructed two classification algorithms—model random forest and support vector classification, for a municipal corporation in a particular city that can label any new water sample as either safe or unsafe for drinking. Which of the two models (random forest or support vector classifier) would we recommend for deployment in the real world?